Add script for lowest high scores
It's pretty easy to look up the players with the lowest high scores
overall, but what about the lowest high score after N innings batted?
That's what this script looks up. The two CTEs can also be used for
other things; for instance, to see players who made it to 100 innings
without reaching 50, we can do:

    SELECT
      innings_count AS innings,
      player,
      cumulative_high_score AS high_score
    FROM running
    WHERE cumulative_high_score < 50 AND innings_count = 100
    ORDER BY cumulative_high_score DESC;

And see a bunch of excellent bowlers (you'd have to be to bat in 100
innings without making many runs!).