Add bowling Bannerman
This is the highest percentage of runs conceded by a bowler in an
innings where the opposition was all out. Unsurprisingly, most of these
happen in very low innings totals. When Nepal Women beat Maldives Women
in the third-place play-off of the 2019 South Asian Women's Games,
Kabita Kunwar conceded 6 of the 8 (!!) runs Maldives Women scored -
all off wides, as there was only a single run off the bat.

This is better than other Bannerman calculations in this repo, because
it uses the actual team score, not just runs off the bat, which means
it's actually accurate. Previously I had not noticed that the
`*_team.csv` files contain the team scores and can easily be joined to
the batting and bowling tables.