Perform fewer Pandas appends
This speeds up the scrapers significantly (on a clean run of the women's
Test team records, it takes two thirds of the time from before) by
changing the appending approach for building up the intermediate
dataframes.

Before, we were constructing a single-row dataframe for each row in the
HTML table. We were then appending those on to a page dataframe (with up
to 200 rows), and finally appending _that_ to the overall dataframe.

This is slow, and it's known to be slow in the Pandas community. See
https://stackoverflow.com/a/56746204 for an example. We can improve this
by replacing all the intermediate dataframes with regular Python lists.
Then, once a page is complete, we construct a single dataframe from that
list of lists, convert it to the correct types, and finally concat that
with the overall dataframe.